+++ src/getEventProps.js
@@ -1,3 +1,11 @@
 const { reactEventprops } = require("./lists/base/react");
 
-module.exports.getEventProps = () => reactEventprops;
+module.exports.getEventProps = arg => {
+  if (arg)
+    //eslint-disable-next-line no-console
+    console.warn(
+      `[react-known-props] getEventProps: Unexpected argument: '${typeof element}' ${arg}.`
+    );
+
+  return reactEventprops;
+};